refactor: Use consistent PascalCase for filenames
